CAMEL-20557: Rest DSL to use openapi spec directly
