+++ components/camel-rest-openapi/src/main/java/org/apache/camel/component/rest/openapi/RestOpenapiProcessorStrategy.java
@@ -20,7 +20,7 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import org.apache.camel.AsyncCallback;
 import org.apache.camel.Exchange;
-import org.apache.camel.Processor;
+import org.apache.camel.support.processor.RestBindingSupport;
 
 /**
  * Strategy for processing the Rest DSL that services an OpenAPI spec.
@@ -64,20 +64,19 @@ public interface RestOpenapiProcessorStrategy {
     /**
      * Strategy for processing the Rest DSL operation
      *
-     * @param  operation     the rest operation
-     * @param  path          the context-path
-     * @param  bindingBefore optional binding to execute before processing the Rest DSL operation
-     * @param  bindingAfter  optional binding to execute after processing the Rest DSL operation
-     * @param  exchange      the exchange
-     * @param  callback      the AsyncCallback will be invoked when the processing of the exchange is completed. If the
-     *                       exchange is completed synchronously, then the callback is also invoked synchronously. The
-     *                       callback should therefore be careful of starting recursive loop.
-     * @return               (doneSync) true to continue execute synchronously, false to continue being executed
-     *                       asynchronously
+     * @param  operation the rest operation
+     * @param  path      the context-path
+     * @param  binding   binding advice
+     * @param  exchange  the exchange
+     * @param  callback  the AsyncCallback will be invoked when the processing of the exchange is completed. If the
+     *                   exchange is completed synchronously, then the callback is also invoked synchronously. The
+     *                   callback should therefore be careful of starting recursive loop.
+     * @return           (doneSync) true to continue execute synchronously, false to continue being executed
+     *                   asynchronously
      */
     boolean process(
             Operation operation, String path,
-            Processor bindingBefore, Processor bindingAfter,
+            RestBindingSupport binding,
             Exchange exchange, AsyncCallback callback);
 
     /**
